lib: tenstorrent: move reg header to shared include
Move reg header to shared include, since it will be used by many
discrete components
