+++ lib/tenstorrent/bh_arc/avs.c
@@ -9,7 +9,7 @@
 #include <zephyr/sys/util.h>
 
 #include "timer.h"
-#include "reg.h"
+#include <tenstorrent/reg.h>
 #include "pll.h"
 
 #define APB2AVSBUS_AVS_INTERRUPT_MASK_REG_ADDR 0x80100034
