+++ lib/tenstorrent/bh_arc/efuse.c
@@ -10,7 +10,7 @@
 
 #include "timer.h"
 #include "dw_apb_i2c.h"
-#include "reg.h"
+#include <tenstorrent/reg.h>
 #include "efuse.h"
 
 #define EFUSE_DFT0_MEM_BASE_ADDR          0x80040000
