+++ lib/tenstorrent/bh_arc/gddr.c
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include <reg.h>
+#include <tenstorrent/reg.h>
 #include <noc.h>
 #include <noc2axi.h>
 #include <arc_dma.h>
